+++ src/parser/inline/autolink/channel.ts
@@ -5,9 +5,16 @@ import { account } from './account';
 import { hashtag } from './hashtag';
 import { define } from 'typed-dom';
 
+// https://example.com/@user?ch=a+b must be a user channel page or a redirect page going there.
+
 export const channel: AutolinkParser.ChannelParser = validate('@', fmap(
   sequence([
     account,
     some(hashtag),
   ]),
-  es => [define(es[0], { class: 'channel' }, stringify(es))]));
+  es => {
+    const source = stringify(es);
+    const el = es[0];
+    const url = `${el.getAttribute('href')}?ch=${source.slice(source.indexOf('#') + 1).replace(/#/g, '+')}`;
+    return [define(el, { class: 'channel', href: url }, source)];
+  }));